+++ sdk/tests/types/transaction_regular_essence.rs
@@ -13,7 +13,10 @@ use iota_sdk::types::block::{
         Output, SimpleTokenScheme, TokenId, TokenScheme,
     },
     payload::{
-        transaction::{RegularTransactionEssence, TransactionId, TransactionPayload},
+        transaction::{
+            RegularTransactionEssence, TransactionCapabilities, TransactionCapabilityFlag, TransactionId,
+            TransactionPayload,
+        },
         Payload,
     },
     protocol::protocol_parameters,
@@ -463,4 +466,47 @@ fn duplicate_output_foundry() {
         essence,
         Err(Error::DuplicateOutputChain(ChainId::Foundry(foundry_id_0))) if foundry_id_0 == foundry_id
     ));
+}
+
+#[test]
+fn transactions_capabilities() {
+    let protocol_parameters = protocol_parameters();
+    let transaction_id = TransactionId::new(prefix_hex::decode(TRANSACTION_ID).unwrap());
+    let input1 = Input::Utxo(UtxoInput::new(transaction_id, 0).unwrap());
+    let input2 = Input::Utxo(UtxoInput::new(transaction_id, 1).unwrap());
+    let address = Address::from(Ed25519Address::new(prefix_hex::decode(ED25519_ADDRESS_1).unwrap()));
+    let amount = 1_000_000;
+    let output = Output::Basic(
+        BasicOutput::build_with_amount(amount)
+            .add_unlock_condition(AddressUnlockCondition::new(address))
+            .finish_with_params(&protocol_parameters)
+            .unwrap(),
+    );
+    let essence = RegularTransactionEssence::builder(protocol_parameters.network_id(), rand_inputs_commitment())
+        .with_inputs(vec![input1, input2])
+        .add_output(output)
+        .add_mana_allotment(rand_mana_allotment(&protocol_parameters))
+        .finish_with_params(&protocol_parameters)
+        .unwrap();
+    let mut capabilities = essence.capabilities().clone();
+
+    use TransactionCapabilityFlag as Flag;
+
+    assert!(capabilities.is_none());
+
+    assert!(!capabilities.has_capability(Flag::BurnNativeTokens));
+    capabilities.add_capability(Flag::BurnNativeTokens);
+    assert!(capabilities.has_capabilities([Flag::BurnNativeTokens]));
+
+    assert!(!capabilities.has_capability(Flag::BurnMana));
+    capabilities.set_capabilities([Flag::BurnMana, Flag::DestroyAccountOutputs]);
+    assert!(capabilities.has_capabilities([Flag::BurnMana, Flag::DestroyAccountOutputs]));
+    assert!(!capabilities.has_capability(Flag::BurnNativeTokens));
+
+    assert!(!capabilities.is_none());
+
+    assert!(!capabilities.has_capabilities(TransactionCapabilities::all().capabilities_iter()));
+    capabilities.set_all();
+    assert!(capabilities.has_capabilities(TransactionCapabilities::all().capabilities_iter()));
+    assert!(capabilities.has_capabilities([Flag::DestroyFoundryOutputs, Flag::DestroyNftOutputs]));
 }
